refactor: server entry structure
