+++ server.js
@@ -1,33 +1,32 @@
 require('dotenv').config();
 
 const express = require('express');
-const mongoose = require('mongoose');
 const { resolve } = require('path');
+const homeRoutes = require('./src/routes/home.routes');
+const userRoutes = require('./src/routes/user.routes');
+const itemRoutes = require('./src/routes/item.routes');
+const connectToDatabase = require('./src/utils/connectToDatabase');
 
-const connectionPort = process.env.PORT || 3000;
+const connectionPort = process.env.PORT || '3000';
 const app = express();
-const homeRoutes = require('./routes/home.routes');
-const userRoutes = require('./routes/user.routes');
-const itemRoutes = require('./routes/item.routes');
 
-mongoose.connect(
-  'mongodb+srv://danielsantos:staging010203@staging1.9jkrt.mongodb.net/staging?retryWrites=true&w=majority',
-  { useNewUrlParser: true, useUnifiedTopology: true },
-)
-  .then(() => app.emit('database connected'))
-  .catch(() => app.emit('database failed'));
+const run = async () => {
+  try {
+    await connectToDatabase();
 
-app.use(express.urlencoded({ extended: true }));
+    app.use(express.urlencoded({ extended: true }));
 
-app.set('views', resolve(__dirname, 'src', 'views'));
-app.set('view engine', 'ejs');
+    app.set('views', resolve(__dirname, 'src', 'views'));
+    app.set('view engine', 'ejs');
 
-app.use(homeRoutes);
-app.use(userRoutes);
-app.use(itemRoutes);
+    app.use(homeRoutes);
+    app.use(userRoutes);
+    app.use(itemRoutes);
 
-app.on('database connected', () => {
-  app.listen(connectionPort, () => console.log(`successfull listening to port ${connectionPort}`));
-});
+    app.listen(connectionPort, () => console.log(`successfull listening to port ${connectionPort}`));
+  } catch (err) {
+    console.log('fail while connecting to database');
+  }
+};
 
-app.on('database failed', () => console.log('fail while connecting to database'));
+run();
