+++ src/modules/database/schema/Item.js
@@ -6,6 +6,7 @@ const ItemSchema = new mongoose.Schema({
   quantity: { type: String, required: true },
   ownerId: { type: String, require: true },
   description: { type: String, require: false },
+  category: { type: String, require: false },
   createdAt: { type: Date, required: true },
   updatedAt: { type: Date, required: false },
 });
