+++ src/routes/home.routes.js
@@ -1,7 +1,7 @@
 const express = require('express');
 
 const router = express.Router();
-const homeController = require('../src/controllers/home.controller');
+const homeController = require('../controllers/home.controller');
 
 router.get('/', homeController.get);
