+++ src/routes/user.routes.js
@@ -1,7 +1,7 @@
 const express = require('express');
 
 const router = express.Router();
-const userController = require('../src/controllers/user.controller');
+const userController = require('../controllers/user.controller');
 
 router.post('/user/create', userController.create);
 router.patch('/user', userController.update);
