+++ src/utils/connectToDatabase.js
@@ -1,0 +1,8 @@
+const mongoose = require('mongoose');
+
+module.exports = async () => {
+  await mongoose.connect(
+    process.env.CONNECTION_STRING,
+    { useNewUrlParser: true, useUnifiedTopology: true },
+  );
+};
